refactor: Deprecated `InstallSuccess` component
